Documentation: Added Javadocs the source files

+++ consumer/src/main/java/com/example/dto/Command.java
@@ -7,7 +7,17 @@ import java.util.UUID;
 @Data
 public class Command {
 
+    /**
+     * Randomly generated id. Format UUID
+     */
     private UUID id;
-    private Long number;
+    /**
+     * Sequence number of the Command
+     */
+    private Integer number;
+    /**
+     * Indicates process time of the Command.
+     * The range is between 3 and 7 seconds
+     */
     private Integer capacity;
 }
